16/08/24

+++ src/sneakers_shop/views.py
@@ -2,16 +2,22 @@ import math
 from datetime import date
 
 import requests
+from django.contrib.auth import get_user_model, login
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.views import LoginView, LogoutView
 from django.db.models import Max, Min, Q
 from django.http import Http404, HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from django.template.defaulttags import register
-from django.views.generic import DetailView, ListView, TemplateView, View
+from django.urls import reverse_lazy
+from django.views.generic import (CreateView, DetailView, ListView,
+                                  TemplateView, View)
 from webargs import fields
 from webargs.djangoparser import use_args
 
-from sneakers_shop.forms import CartAddForm
-from sneakers_shop.models import Carts, Sneakers
+from sneakers_shop.forms import (CartAddForm, UserLoginForm,
+                                 UserRegistrationForm)
+from sneakers_shop.models import Carts, Sneakers, Wishlists
 
 
 class IndexView(TemplateView):
@@ -74,6 +80,16 @@ class ShopListView(ListView):
         sorted_by = params.get("sort")
         min_price = self.request.POST.get("min_price")
         max_price = self.request.POST.get("max_price")
+        search = self.request.POST.get("search")
+
+        if search:
+            sneakers = Sneakers.objects.filter(
+                Q(size_sneakers__icontains=search)
+                | Q(color_sneakers__icontains=search)  # NOQA W503
+                | Q(model_sneakers__icontains=search)  # NOQA W503
+                | Q(brand_sneakers__icontains=search)  # NOQA W503
+            )
+            return sneakers
 
         if min_price and max_price:
             params["min_price"] = min_price
@@ -92,6 +108,10 @@ class ShopListView(ListView):
             self.request.session[f"{param_name}"] = filters[param_name]
 
             for k, v in self.request.session.items():
+                if k == "wishlist":
+                    continue
+                if k.startswith("_"):
+                    continue
                 if k == "min_price":
                     or_filter &= Q(price_sneakers__gte=v)
                     continue
@@ -128,12 +148,12 @@ class ShopListView(ListView):
             user_min_price = db_min_price
             user_max_price = db_max_price
 
-        for shoe in Sneakers.objects.all():
-            brand = shoe.brand_sneakers
-            unique_brands.add(shoe.brand_sneakers)
-            unique_models.add(shoe.model_sneakers)
-            unique_color.add(shoe.color_sneakers)
-            unique_sizes.add(shoe.size_sneakers)
+        for shoes in Sneakers.objects.all():
+            brand = shoes.brand_sneakers
+            unique_brands.add(shoes.brand_sneakers)
+            unique_models.add(shoes.model_sneakers)
+            unique_color.add(shoes.color_sneakers)
+            unique_sizes.add(shoes.size_sneakers)
             if brand in brands_count:
                 brands_count[brand] += 1
             else:
@@ -156,6 +176,19 @@ class ShopListView(ListView):
         return context
 
 
+class SneakersDetailView(DetailView):
+    model = Sneakers
+    context_object_name = "sneakers"
+    template_name = "shop-detail.html"
+
+    def get_queryset(self):
+        queryset = Sneakers.objects.filter(id=self.kwargs["pk"])
+        if queryset.exists():
+            return Sneakers.objects.filter(id=self.kwargs["pk"])
+        else:
+            raise Http404
+
+
 class CartListView(ListView):
     model = Carts
     template_name = "cart.html"
@@ -163,10 +196,13 @@ class CartListView(ListView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
+
         if self.request.user.is_authenticated:
             carts = Carts.objects.filter(user=self.request.user)
-            context["carts"] = Carts.objects.filter(user=self.request.user)
+            context["carts"] = carts
             total_price = sum(cart.total_price for cart in carts)
+            context["total_price"] = total_price
+
         else:
             cart = self.request.session.get("cart", {})
             context["carts"] = cart.values()
@@ -174,6 +210,7 @@ class CartListView(ListView):
 
         context["form"] = CartAddForm()
         context["total_price"] = total_price
+        print(context["carts"])
         return context
 
 
@@ -181,6 +218,7 @@ class CartAddView(View):
     def get(self, request, *args, **kwargs):
         product_id = kwargs.get("pk")
         product = get_object_or_404(Sneakers, pk=product_id)
+
         if request.user.is_authenticated:
             cart_item, created = Carts.objects.get_or_create(
                 user=request.user,
@@ -211,7 +249,6 @@ class CartAddView(View):
                     "price": product.price_sneakers,
                 }
             request.session["cart"] = cart
-            print(cart)
         return redirect("cart")
 
     def post(self, request, *args, **kwargs):
@@ -219,8 +256,11 @@ class CartAddView(View):
         product_id = kwargs.get("pk")
         if action == "remove":
             if request.user.is_authenticated:
-                cart_item = get_object_or_404(Carts, user=request.user, sneakers_id=product_id)
-                cart_item.delete()
+                cart_item = Carts.objects.filter(user=request.user, sneakers_id=product_id)
+                if cart_item:
+                    cart_item.delete()
+                else:
+                    pass
             else:
                 cart = request.session.get("cart", {})
                 if str(product_id) in cart:
@@ -229,14 +269,66 @@ class CartAddView(View):
         return redirect("cart")
 
 
-class SneakersDetailView(DetailView):
-    model = Sneakers
-    context_object_name = "sneakers"
-    template_name = "shop-detail.html"
+class UserRegistrationView(CreateView):
+    template_name = "sign_up.html"
+    form_class = UserRegistrationForm
+    success_url = reverse_lazy("index")
 
-    def get_queryset(self):
-        queryset = Sneakers.objects.filter(id=self.kwargs["pk"])
-        if queryset.exists():
-            return Sneakers.objects.filter(id=self.kwargs["pk"])
+    def form_valid(self, form):
+        response = super().form_valid(form)
+        user = self.object
+        login(self.request, user)
+        return response
+
+
+class UserLoginView(LoginView):
+    template_name = "sign_in.html"
+    form_class = UserLoginForm
+    success_url = reverse_lazy("index")
+
+    def form_valid(self, form):
+        return super().form_valid(form)
+
+
+class UserLogoutView(LogoutView):
+    next_page = reverse_lazy("index")
+
+
+class WishlistAddView(View):
+    def get(self, request, *args, **kwargs):
+        product_id = kwargs.get("pk")
+        product = get_object_or_404(Sneakers, pk=product_id)
+
+        if request.user.is_authenticated:
+            Wishlists.objects.get_or_create(
+                user=request.user,
+                product=product,
+            )
         else:
-            raise Http404
+            wishlist = request.session.get("wishlist", {})
+            wishlist[str(product_id)] = {
+                "product_id": product_id,
+                "model_sneakers": product.model_sneakers,
+                "brand_sneakers": product.brand_sneakers,
+                "image_sneakers": product.image_sneakers.url if product.image_sneakers else None,
+                "price": product.price_sneakers,
+            }
+            request.session["wishlist"] = wishlist
+        return redirect("wishlist")
+
+
+class WishListView(ListView):
+    model = Wishlists
+    template_name = "wishlist.html"
+    context_object_name = "wishlists"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        if self.request.user.is_authenticated:
+            context["wishlists"] = Wishlists.objects.filter(user=self.request.user)
+        else:
+            wishlist = self.request.session.get("wishlist", {})
+            context["wishlist"] = wishlist.values()
+
+        return context
